FIX: Skip dirs without id or name instead of aborting
There are directories now that contain neither a name or an id
delivered by the API for some reason, looking like this:
{
    'is_visible': True,
    'is_readable': False,
    'is_writable': True
}
which makes no sense at all. If such a directory is given the
processing is stopped here and all further directories are ignored.
This should not be the case.

This commit fixes this behavior, so that all directories are processed
as expected.